Refactor: contract router (#1062)
