+++ backend/src/database/queries.ts
@@ -995,84 +995,6 @@ export const queryExecutedReceiptsList = async (blockHash: string) => {
     .execute();
 };
 
-export const queryContractInfo = async (accountId: string) => {
-  // find the latest update in analytics db
-  const latestUpdateResult = await analyticsDatabase
-    .selectFrom("deployed_contracts")
-    .select("deployed_at_block_timestamp as latest_updated_timestamp")
-    .orderBy("deployed_at_block_timestamp", "desc")
-    .limit(1)
-    .executeTakeFirstOrThrow();
-  const { latest_updated_timestamp: latestUpdatedTimestamp } =
-    latestUpdateResult;
-  // query for the latest info from indexer
-  // if it return 'undefined' then there was no update since deployed_at_block_timestamp
-  const contractInfoFromIndexer = await indexerDatabase
-    .selectFrom("action_receipt_actions")
-    .leftJoin("receipts", (jb) =>
-      jb.onRef("receipts.receipt_id", "=", "action_receipt_actions.receipt_id")
-    )
-    .select([
-      (eb) =>
-        div(
-          eb,
-          "receipts.included_in_block_timestamp",
-          1000 * 1000,
-          "block_timestamp"
-        ),
-      "receipts.originated_from_transaction_hash as hash",
-    ])
-    .where("action_kind", "=", "DEPLOY_CONTRACT")
-    .where("receipt_receiver_account_id", "=", accountId)
-    .where("receipt_included_in_block_timestamp", ">", latestUpdatedTimestamp)
-    .orderBy("receipt_included_in_block_timestamp", "desc")
-    .limit(1)
-    .executeTakeFirst();
-
-  if (contractInfoFromIndexer) {
-    return {
-      block_timestamp: contractInfoFromIndexer.block_timestamp,
-      // TODO: Discover how to get rid of non-null type assertion
-      hash: contractInfoFromIndexer.hash!,
-    };
-  }
-
-  // query to analytics db to find latest historical record
-  const contractInfoFromAnalytics = await analyticsDatabase
-    .selectFrom("deployed_contracts")
-    .select([
-      "deployed_by_receipt_id as receipt_id",
-      (eb) =>
-        div(eb, "deployed_at_block_timestamp", 1000 * 1000, "block_timestamp"),
-    ])
-    .where("deployed_to_account_id", "=", accountId)
-    .orderBy("deployed_at_block_timestamp", "desc")
-    .limit(1)
-    .executeTakeFirstOrThrow();
-
-  // query for transaction hash where contact was deployed
-  const transactionHashResult = await indexerDatabase
-    .selectFrom("receipts")
-    .select("originated_from_transaction_hash as hash")
-    .where(
-      "receipt_id",
-      "=",
-      // Different "flavors" in different DBs
-      // there is no way to fix types but force casting
-      contractInfoFromAnalytics.receipt_id as Indexer.ReceiptsId
-    )
-    .limit(1)
-    .executeTakeFirstOrThrow();
-
-  if (contractInfoFromAnalytics) {
-    return {
-      block_timestamp: contractInfoFromAnalytics.block_timestamp,
-      hash: transactionHashResult.hash,
-    };
-  }
-  return undefined;
-};
-
 // chunks
 export const queryGasUsedInChunks = async (blockHash: string) => {
   return indexerDatabase
