+++ backend/src/router/contract/index.ts
@@ -1,0 +1,5 @@
+import * as trpc from "@trpc/server";
+import { Context } from "../../context";
+import { router as byIdRouter } from "./by-id";
+
+export const router = trpc.router<Context>().merge(byIdRouter);
